requirements.txt
aws-cdk.core>=1.15.0
aws-cdk.aws_ec2>=1.15.0
aws_cdk.aws_ecs>=1.15.0
aws_cdk.aws_s3>=1.15.0
aws_cdk.aws_route53>=1.15.0
aws-cdk.aws-rds>=1.15.0
